Rewrite interactable objects, Daily Spin Tent/Wheel
Rewrote how interactable objects are handled. Used to be just for NPCs where players would have to walk up to the object (NPC) in order for the interaction to go through but is now written with callbacks and includes all objects that should be interacted with in that fashion instead of just NPCs.

Daily Spin Tent implemented as well as the interactable Daily Spin Wheel so that the Daily Spin game is now fully implemented.